+++ src/__internal__/input/input.spec.tsx
@@ -5,7 +5,7 @@ import Input, { InputProps, EnterKeyHintTypes } from "./input.component";
 import StyledInput from "./input.style";
 
 import { InputContext, InputContextProps } from "../input-behaviour";
-import { SelectTextboxContext } from "../../components/select/select-textbox/select-textbox-context";
+import { SelectTextboxContext } from "../../components/select/__internal__/select-textbox/select-textbox.context";
 
 import { assertStyleMatch } from "../../__spec_helper__/__internal__/test-utils";
