+++ src/components/select/__internal__/list-action-button/list-action-button.style.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import StyledButton from "../../button/button.style";
+import StyledButton from "../../../button/button.style";
 
 const StyledListActionButtonWrapper = styled.div`
   padding-top: var(--spacing100);
