+++ src/components/select/__internal__/select-textbox/select-textbox.component.tsx
@@ -1,13 +1,13 @@
 import React, { useMemo } from "react";
 import { offset, size as sizeMiddleware } from "@floating-ui/dom";
 
-import useFloating from "../../../hooks/__internal__/useFloating";
-import Textbox, { CommonTextboxProps } from "../../textbox";
-import SelectText from "../__internal__/select-text";
-import useLocale from "../../../hooks/__internal__/useLocale";
-import { ValidationProps } from "../../../__internal__/validations";
-import { CustomSelectChangeEvent } from "../simple-select/simple-select.component";
-import { SelectTextboxContext } from "./select-textbox-context";
+import useFloating from "../../../../hooks/__internal__/useFloating";
+import Textbox, { CommonTextboxProps } from "../../../textbox";
+import SelectText from "../select-text";
+import useLocale from "../../../../hooks/__internal__/useLocale";
+import { ValidationProps } from "../../../../__internal__/validations";
+import { CustomSelectChangeEvent } from "../../simple-select/simple-select.component";
+import { SelectTextboxContext } from "./select-textbox.context";
 
 const floatingMiddleware = [
   offset(({ rects }) => ({
