+++ src/components/select/__internal__/utils/get-next-child-by-text.ts
