+++ src/components/select/__internal__/utils/get-next-index-by-key.ts
