+++ src/components/select/__internal__/utils/highlight-part-of-text.tsx
