+++ src/components/select/__internal__/utils/is-expected-option.spec.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Option from "../option";
+import Option from "../../option";
 import isExpectedOption from "./is-expected-option";
 
 describe("isExpectedOption", () => {
