+++ src/components/select/__internal__/utils/is-expected-value.spec.ts
