+++ src/components/select/__internal__/utils/is-navigation-key.spec.ts
