+++ src/components/select/filterable-select/filterable-select.mdx
@@ -2,7 +2,7 @@ import { Meta, ArgTypes, Canvas } from "@storybook/blocks";
 import TranslationKeysTable from "../../../../.storybook/utils/translation-keys-table";
 
 import * as OptionStories from "../option/option.stories";
-import * as SelectTextboxStories from "../select-textbox/select-textbox.stories";
+import * as SelectTextboxStories from "../__internal__/select-textbox/select-textbox.stories";
 import * as FilterableSelectStories from "./filterable-select.stories";
 
 <Meta title="Select/Filterable" of={FilterableSelectStories} />
