+++ src/components/select/filterable-select/filterable-select.spec.tsx
@@ -16,12 +16,12 @@ import {
 import { FilterableSelect, Option, FilterableSelectProps } from "..";
 import StyledOption from "../option/option.style";
 import Textbox from "../../textbox";
-import MatchingText from "../utils/matching-text.style";
-import SelectList from "../select-list/select-list.component";
+import MatchingText from "../__internal__/utils/matching-text.style";
+import SelectList from "../__internal__/select-list/select-list.component";
 import {
   StyledSelectListContainer,
   StyledScrollableContainer,
-} from "../select-list/select-list.style";
+} from "../__internal__/select-list/select-list.style";
 import Button from "../../button";
 import Label from "../../../__internal__/label";
 import InputIconToggle from "../../../__internal__/input-icon-toggle";
