+++ src/components/select/multi-select/multi-select.mdx
@@ -2,7 +2,7 @@ import { Meta, ArgTypes, Canvas } from "@storybook/blocks";
 import TranslationKeysTable from "../../../../.storybook/utils/translation-keys-table";
 
 import * as OptionStories from "../option/option.stories";
-import * as SelectTextboxStories from "../select-textbox/select-textbox.stories";
+import * as SelectTextboxStories from "../__internal__/select-textbox/select-textbox.stories";
 import * as MultiSelectStories from "./multi-select.stories";
 
 <Meta title="Select/MultiSelect" of={MultiSelectStories} />
